create a util packages for more-abstract datatypes
